In coda all'aggiornamento viene lanciata la console con i parametri per farle aggiornare tutte le company

+++ Mago4Butler.BL/BL/FileSystemService.cs
@@ -23,19 +23,34 @@ namespace Microarea.Mago4Butler.BL
             DeleteDirectory(instanceRootFolder);
         }
 
-        static void DeleteDirectory(DirectoryInfo dirInfo)
+        public void RemoveAllFilesButTheCustomAndAppDataFolder(Instance instance)
+        {
+            var instanceRootFolder = new DirectoryInfo(Path.Combine(this.rootFolder, instance.Name));
+
+            DeleteDirectory(instanceRootFolder, "Custom", "AppData");
+        }
+
+        static void DeleteDirectory(DirectoryInfo dirInfo, params string[] tokensToBeSkipped)
         {
             var files = dirInfo.GetFiles();
             var subDirs = dirInfo.GetDirectories();
 
             foreach (var file in files)
             {
+                if (tokensToBeSkipped.Contains(file.Name, StringComparer.InvariantCultureIgnoreCase))
+                {
+                    continue;
+                }
                 file.Attributes |= FileAttributes.Normal;
                 File.Delete(file.FullName);
             }
 
             foreach (var subDir in subDirs)
             {
+                if (tokensToBeSkipped.Contains(subDir.Name, StringComparer.InvariantCultureIgnoreCase))
+                {
+                    continue;
+                }
                 DeleteDirectory(subDir);
             }
